refactor(plan_snapshots): move RDA and Extension serializers
Move plan snapshot serializers into the PlanSnapshots namespace to better reflect their purpose and prepare for upcoming snapshot PDF rendering work.

Update serializer specs and PlanSnapshot callers to use the new locations. No behavior changes are introduced.
